docs/piyut-talis-agent/requirements.txt
# Python dependencies for Piyut-Talis-Agent DOCX export functionality
python-docx>=1.0.0
